Challenge 17 was interesting.
The biggest take away was that just because my math is right, doesn't mean it is representative of a valid solution. I ran into an issue where many blocks solved for the oracle. In my math, I assumed that only one oracle solution existed where as my code had to evaluate multiple solutions to the oracle.